LPD-51036 Move files and update imports
