+++ modules/apps/site/site-cms-site-initializer/test/js/main/components/SpaceSticker.test.tsx
@@ -7,7 +7,7 @@ import '@testing-library/jest-dom/extend-expect';
 import {render} from '@testing-library/react';
 import React from 'react';
 
-import SpaceSticker from '../../../src/main/resources/META-INF/resources/js/components/SpaceSticker';
+import SpaceSticker from '../../../../src/main/resources/META-INF/resources/js/main/components/SpaceSticker';
 
 const spaceTitle = 'First space';
